Make sure to purge open-vm-tools before installing
This is needed on Debian, because if the pkg is just deinstalled, the config
files are still present and the vmware-tools installer want to its job.

Basically he does install the kernel modules, but no CLI tools at all.

This will keep the puppet module reinstalling the tools in a cycle.

In addition added ordering to ensure dependencies being checked before
unpacking or installing.